Update type builder macros
Create a dedicated macro to create postgres types, and update t-digest
to use it.
