+++ .gitignore
@@ -4,5 +4,4 @@
 /target
 *.iml
 **/*.rs.bk
-Cargo.lock
 sql/*.generated.sql
